# feat: skip filling NULL for put and delete requests