--- a/src/mito2/src/memtable/key_values.rs
+++ b/src/mito2/src/memtable/key_values.rs
@@ -28,7 +28,7 @@ pub struct KeyValues {
     /// must not be `None`.
     mutation: Mutation,
     /// Key value read helper.
-    helper: ReadRowHelper,
+    helper: SparseReadRowHelper,
 }
 
 impl KeyValues {
@@ -37,7 +37,18 @@ impl KeyValues {
     /// Returns `None` if `rows` of the `mutation` is `None`.
     pub fn new(metadata: &RegionMetadata, mutation: Mutation) -> Option<KeyValues> {
         let rows = mutation.rows.as_ref()?;
-        let helper = ReadRowHelper::new(metadata, rows);
+        let helper = SparseReadRowHelper::new(metadata, rows);
+
+        Some(KeyValues { mutation, helper })
+    }
+
+    /// Creates [KeyValues] from specific `mutation`. This method is used for [Mutation] that
+    /// has not been filled with all columns.
+    ///
+    /// Returns `None` if `rows` of the `mutation` is `None`.
+    pub fn new_sparse(metadata: &RegionMetadata, mutation: Mutation) -> Option<KeyValues> {
+        let rows = mutation.rows.as_ref()?;
+        let helper = SparseReadRowHelper::new(metadata, rows);
 
         Some(KeyValues { mutation, helper })
     }
@@ -75,7 +86,7 @@ impl KeyValues {
 pub struct KeyValue<'a> {
     row: &'a Row,
     schema: &'a Vec<ColumnSchema>,
-    helper: &'a ReadRowHelper,
+    helper: &'a SparseReadRowHelper,
     sequence: SequenceNumber,
     op_type: OpType,
 }
@@ -85,11 +96,12 @@ impl<'a> KeyValue<'a> {
     pub fn primary_keys(&self) -> impl Iterator<Item = ValueRef> {
         self.helper.indices[..self.helper.num_primary_key_column]
             .iter()
-            .map(|idx| {
-                api::helper::pb_value_to_value_ref(
-                    &self.row.values[*idx],
-                    &self.schema[*idx].datatype_extension,
-                )
+            .map(|idx| match idx {
+                Some(i) => api::helper::pb_value_to_value_ref(
+                    &self.row.values[*i],
+                    &self.schema[*i].datatype_extension,
+                ),
+                None => ValueRef::Null,
             })
     }
 
@@ -97,18 +109,19 @@ impl<'a> KeyValue<'a> {
     pub fn fields(&self) -> impl Iterator<Item = ValueRef> {
         self.helper.indices[self.helper.num_primary_key_column + 1..]
             .iter()
-            .map(|idx| {
-                api::helper::pb_value_to_value_ref(
-                    &self.row.values[*idx],
-                    &self.schema[*idx].datatype_extension,
-                )
+            .map(|idx| match idx {
+                Some(i) => api::helper::pb_value_to_value_ref(
+                    &self.row.values[*i],
+                    &self.schema[*i].datatype_extension,
+                ),
+                None => ValueRef::Null,
             })
     }
 
     /// Get timestamp.
     pub fn timestamp(&self) -> ValueRef {
         // Timestamp is primitive, we clone it.
-        let index = self.helper.indices[self.helper.num_primary_key_column];
+        let index = self.helper.indices[self.helper.num_primary_key_column].unwrap();
         api::helper::pb_value_to_value_ref(
             &self.row.values[index],
             &self.schema[index].datatype_extension,
@@ -192,6 +205,61 @@ impl ReadRowHelper {
         }
 
         ReadRowHelper {
+            indices,
+            num_primary_key_column: metadata.primary_key.len(),
+        }
+    }
+}
+
+/// Helper to read rows in key, value order for sparse data.
+#[derive(Debug)]
+struct SparseReadRowHelper {
+    /// Key and value column indices.
+    ///
+    /// `indices[..num_primary_key_column]` are primary key columns, `indices[num_primary_key_column]`
+    /// is the timestamp column and remainings are field columns.
+    indices: Vec<Option<usize>>,
+    /// Number of primary key columns.
+    num_primary_key_column: usize,
+}
+
+impl SparseReadRowHelper {
+    /// Creates a [SparseReadRowHelper] for specific `rows`.
+    ///
+    /// # Panics
+    /// Time index column must exist.
+    fn new(metadata: &RegionMetadata, rows: &Rows) -> SparseReadRowHelper {
+        // Build a name to index mapping for rows.
+        let name_to_index: HashMap<_, _> = rows
+            .schema
+            .iter()
+            .enumerate()
+            .map(|(index, col)| (&col.column_name, index))
+            .collect();
+        let mut indices = Vec::with_capacity(metadata.column_metadatas.len());
+
+        // Get primary key indices.
+        for pk_column_id in &metadata.primary_key {
+            // Safety: Id comes from primary key.
+            let column = metadata.column_by_id(*pk_column_id).unwrap();
+            let index = name_to_index.get(&column.column_schema.name);
+            indices.push(index.copied());
+        }
+        // Get timestamp index.
+        // Safety: time index must exist
+        let ts_index = name_to_index
+            .get(&metadata.time_index_column().column_schema.name)
+            .unwrap();
+        indices.push(Some(*ts_index));
+
+        // Iterate columns and find field columns.
+        for column in metadata.field_columns() {
+            // Get index in request for each field column.
+            let index = name_to_index.get(&column.column_schema.name);
+            indices.push(index.copied());
+        }
+
+        SparseReadRowHelper {
             indices,
             num_primary_key_column: metadata.primary_key.len(),
         }