I hereby agree to the terms of the [GreptimeDB CLA](https://gist.github.com/xtang/6378857777706e568c1949c7578592cc)

## What's changed and what's your intention?


As the title. Test result:

- Cut put latency by half

![image](https://github.com/GreptimeTeam/greptimedb/assets/15380403/c339e42c-899b-4caf-8a8b-a2853abb5231)

- Double put throughput (18w to 42w)


| Before | After |
|--------|--------|
| ![image](https://github.com/GreptimeTeam/greptimedb/assets/15380403/37c1dbe0-c4f1-4348-b98a-850c96a6cfc1) | ![image](https://github.com/GreptimeTeam/greptimedb/assets/15380403/fda37378-2a36-44cc-b809-7d574d19896e) | 

- No regression in TSBS (and somehow has a little improvement...)

Before
```
Summary:
loaded 1036800000 metrics in 190.468sec with 6 workers (mean rate 5443439.71 metrics/sec)
loaded 103680000 rows in 190.468sec with 6 workers (mean rate 544343.97 rows/sec)
```

After
```
Summary:
loaded 1036800000 metrics in 187.196sec with 6 workers (mean rate 5538568.26 metrics/sec)
loaded 103680000 rows in 187.196sec with 6 workers (mean rate 553856.83 rows/sec)
```

## Checklist

- [ ]  I have written the necessary rustdoc comments.
- [ ]  I have added the necessary unit tests and integration tests.
- [x]  This PR does not require documentation updates.

## Refer to a related PR or issue link (optional)